+++ src/io/Instance.h
@@ -10,8 +10,8 @@ public:
     int m; // Customers
     
     // Arrays to match the problem data
-    vector<double> opening_costs;          // Cost to open facility i
-    vector<vector<double>> allocation_costs; // Cost to serve client j from facility i
+    vector<long> opening_costs;          // Cost to open facility i
+    vector<vector<long>> allocation_costs; // Cost to serve client j from facility i
     
     string filePath;
